requirements.txt
black==24.3.0
beartype==0.18.0
flake8==7.0.0
h5py==3.10.0
hydra-core==1.3.2
imageio==2.34.0
ipykernel==6.29.3
ipython==8.22.2
isort==5.13.2
jaxtyping==0.2.28
matplotlib==3.8.3
moviepy==1.0.3
numpy==1.26.4
omegaconf==2.3.0
opencv-python==4.9.0.80
pandas==2.2.1
POT==0.9.3
pre-commit==3.7.0
pytictac==1.0.9
scipy==1.12.0
seaborn==0.13.2
tensordict==0.3.1
torchrl==0.3.1
tqdm==4.66.2
virtualenv==20.25.1
wandb==0.17.8
gymnasium==0.29.1
gymnasium-robotics==1.2.4
einops==0.7.0
minari==0.4.3
zarr==2.18.2